# Clean up models.py