--- a/odoo/addons/base/tests/__init__.py
+++ b/odoo/addons/base/tests/__init__.py
@@ -7,6 +7,7 @@ from . import test_api
 from . import test_base
 from . import test_basecase
 from . import test_cache
+from . import test_deprecation
 from . import test_db_cursor
 from . import test_expression
 from . import test_float