--- a/odoo/addons/base/tests/test_deprecation.py
+++ b/odoo/addons/base/tests/test_deprecation.py
@@ -7,6 +7,8 @@ from odoo.tests.common import TransactionCase, tagged
 DEPRECATED_MODEL_ATTRIBUTES = [
     'view_init',
     '_needaction',
+    '_sql',
+    '_execute_sql',
 ]