[REF] base: replace `_browse` by `__init__`

The creation of recordset object was done by `_browse` instead
of a normal python class with the `__init__`.
We remove the old usage of `__init__`, we can now reuse it
with a normal usage.
After this commit, we can create recordset by the Class without call a
classmethod: `self.env['model_name'](env, ids, prefetch_ids)`

[REM] base: depreciate _execute_sql

It wasn't unused then depreciate to be removed in the next version.

[REM] base: remove deadcode from models.py

- Remove unused imports `AsIs` and `Collector`
- Remove _schema variable (unused)
- Remove same_name function (unused)
- Remove attribute `_needaction` which is useless
- Remove backward compatibility `__new__` and `__init__` which isn't
used anymore.
- Remove backward compatibility `__export_rows`

[FIX] base: two small optimisation

- The `update` method of BaseModel wasn't batch for no
reason. This method is used in `_onchange_eval` and in few
onchange in Odoo.
- In `_modified_triggers` avoid a useless record union.

[REM] base: depreciated refresh method

The `refresh` method of models.py is a duplicate of
`invalidate_cache` and deprecated since 8.0 but without any
warning. Add this warning to be able to completely remove it
in the next version.

[REM] base: remove useless view_init

view_init is a useless hook call from default_get.
We can get the same result by overriding default_get
directly.