@rco-odoo 2021-11-15
> You may use `type()` instead of a direct call to the `__class__` dunder if you wanna stick with pep8 recommendations.

From what I could measure, there is no significant performance difference between `type(x)` and `x.__class__`.