@ryv-odoo 2021-11-15
Yep, `type` seems to be the good way to do it. I did it like that because `type(self)(self.env, (id_,)` sound weird to me.
I am going to make the change. Thank you @rco-odoo  and @Julien00859.