@xmo-odoo 2021-11-15
> > You may use `type()` instead of a direct call to the `__class__` dunder if you wanna stick with pep8 recommendations.
> 
> From what I could measure, there is no significant performance difference between `type(x)` and `x.__class__`.

```
❯ python -mtimeit -s 'val = 1' 'type(val)'
5000000 loops, best of 5: 50.4 nsec per loop
❯ python -mtimeit -s 'val = 1' 'val.__class__'
10000000 loops, best of 5: 22.6 nsec per loop
```
as `browse` is called *a lot`, way back then we made the conscious decision to make it as fast as we could, at some flexibibility / prettiness costs.

Things have actually gone worse over time, in 2.7 `type(val)` takes only 40ns.