@rco-odoo 2021-11-15
@xmo-odoo for the record, I did my own test with a record in an Odoo shell:
```python
Python 3.8.10 (default, Sep 28 2021, 16:10:42)
Type 'copyright', 'credits' or 'license' for more information
IPython 7.13.0 -- An enhanced Interactive Python. Type '?' for help.

In [1]: self
Out[1]: res.users(1,)

In [2]: %timeit type(self)
41.4 ns ± 0.0387 ns per loop (mean ± std. dev. of 7 runs, 10000000 loops each)

In [3]: %timeit self.__class__
40.5 ns ± 0.162 ns per loop (mean ± std. dev. of 7 runs, 10000000 loops each)
```