@xmo-odoo 2021-11-16
> @xmo-odoo for the record, I did my own test with a record in an Odoo shell:
> 
> ```python
> Python 3.8.10 (default, Sep 28 2021, 16:10:42)
> Type 'copyright', 'credits' or 'license' for more information
> IPython 7.13.0 -- An enhanced Interactive Python. Type '?' for help.
> 
> In [1]: self
> Out[1]: res.users(1,)
> 
> In [2]: %timeit type(self)
> 41.4 ns ± 0.0387 ns per loop (mean ± std. dev. of 7 runs, 10000000 loops each)
> 
> In [3]: %timeit self.__class__
> 40.5 ns ± 0.162 ns per loop (mean ± std. dev. of 7 runs, 10000000 loops each)
> ```

Might be an ipython thing? Because using `timeit` in a normal python shell I get way more of a difference:
```python
>>> timeit.repeat('type(self)', globals={'self': self}, number=10000000)
[0.6831406939891167, 0.6632596859999467, 0.6566481170011684, 0.6589524900045944, 0.6558081770053832]
>>> timeit.repeat('self.__class__', globals={'self': self}, number=10000000)
[0.4181992550002178, 0.3958683770033531, 0.3977733130013803, 0.3960749170073541, 0.397333968008752]
```
and a native `int` is even more affected:
```python
>>> timeit.repeat('type(self)', globals={'self': 1}, number=10000000)
[0.6810075410030549, 0.6580196689901641, 0.6581342490098905, 0.6559818929963512, 0.6573278839932755]
>>> timeit.repeat('self.__class__', globals={'self': 1}, number=10000000)
[0.3243435699987458, 0.2995552089996636, 0.29956951799977105, 0.30050473200390115, 0.29986495499906596]
```