@ryv-odoo 2021-11-16
Thank you for your comments. 
About the performance between `type` vs `__class__`, I tested too, my result is between yours (Python 3.9.7) :
```
In [3]: timeit.repeat('self.type_perf()', globals={'self': self}, number=10000000)
Out[3]: 
[0.9730675809998957,
 0.9787399080000796,
 1.010729850999951,
 0.9662109689998033,
 0.9469283500000074]

In [4]: timeit.repeat('self.class_perf()', globals={'self': self}, number=10000000)
Out[4]: 
[0.9258530849999715,
 0.8895131269998728,
 0.8908900950000316,
 0.8946078129999933,
 0.8892976620002173]
``` 
I put the code inside a method of the class (as it is used) (but I lose some precision because of the extra call). 

If you look in the bytecode translation: 
```
In [9]: dis.dis(self.class_perf)
5114           0 LOAD_FAST                0 (self)
               2 LOAD_ATTR                0 (__class__)
               4 POP_TOP
               6 LOAD_CONST               0 (None)
               8 RETURN_VALUE

In [10]: dis.dis(self.type_perf)
5111           0 LOAD_GLOBAL              0 (type)
               2 LOAD_FAST                0 (self)
               4 CALL_FUNCTION            1
               6 POP_TOP
               8 LOAD_CONST               0 (None)
              10 RETURN_VALUE

```
I suppose that there is an extra cost from LOAD_GLOBAL and CALL_FUNCTION (which I suppose will read the __class__ attribute inside, right?). Then using __class__ should be faster.

About the change `__init__` vs `_browse`:

> I find the interface change highly debatable, as it makes the "unsafe" browse variant (the one you have to be very careful with) > into the python-standard instantiation method.
> From a UX perspective that seems like the opposite of what we want.

I see what you mean, and now, I am Not confident about this change anymore. `_browse` is more explicit that it is private, and it shouldn't outside the ORM. Then I propose to remove this last commit for now, and maybe we can discuss it at the office and continue with the other changes ? 

@rco-odoo, @xmo-odoo