@rco-odoo 2021-11-16
By the way, I just tested `self.browse()` in an Odoo shell (Python 3.8.10), and the results are interesting:
```python
# before, using class method _browse()
In [1]: %timeit self.browse()
361 ns ± 0.784 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)

# new version, using self.__class__
In [1]: %timeit self.browse()
332 ns ± 1.02 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)

# new version, using type(self)
In [1]: %timeit self.browse()
339 ns ± 0.334 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)
```
On this basic benchmark, the newer versions are slightly faster.

What about keeping the class method `_browse()`, deprecate it and make it call the class?
```python
@classmethod
def _browse(cls, env, ids, prefetch_ids):
    # deprecation warning here
    return cls(env, ids, prefetch_ids)
```

About the API change, my opinion is that this proposal removes technical debt.  The method `_browse()` had no other purpose than working around former technical limitations.  Our CTO pushes for removing such things, even if it breaks third-party code.