@rco-odoo 2021-11-16
Just some extra benchmarks, using your method:
```python
# using class method _browse()
In [2]: timeit.repeat('self.browse()', globals={'self': self}, number=10000000)
Out[2]: 
[3.615988595993258,
 3.595141308993334,
 3.5915285800001584,
 3.5903477589890826,
 3.5900132170063443]

# using self.__class__
In [2]: timeit.repeat('self.browse()', globals={'self': self}, number=10000000)
Out[2]: 
[3.326110805006465,
 3.3221645090088714,
 3.3154461740050465,
 3.316377871000441,
 3.3157517370127607]

# using type(self)
In [2]: timeit.repeat('self.browse()', globals={'self': self}, number=10000000)
Out[2]: 
[3.371973495988641,
 3.3592823880026117,
 3.3603266020072624,
 3.358374931005528,
 3.3597194670001045]
```