@xmo-odoo 2021-11-16
> ```python
> # before, using class method _browse()
> In [1]: %timeit self.browse()
> 361 ns ± 0.784 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)
> 
> # new version, using self.__class__
> In [1]: %timeit self.browse()
> 332 ns ± 1.02 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)
> 
> # new version, using type(self)
> In [1]: %timeit self.browse()
> 339 ns ± 0.334 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)
> ```
> 
> On this basic benchmark, the newer versions are slightly faster.

The "newer versions" here are depending whether `self.browse()` ends with `return self._browse(...)`, `return self.__class__(...)` or `return type(self)(...)`?